Unify tag and title edit modals
